Finish web UI and add auth to appropriate endpoints (#9)

+++ lambdas/delete-ip-address/index.js
@@ -2,18 +2,22 @@
 
 const {
   DynamoDBClient,
+  GetItemCommand,
   DeleteItemCommand,
 } = require("@aws-sdk/client-dynamodb");
 
 exports.handler = async (event) => {
+  const userId = event.requestContext.authorizer.claims.sub;
+
+  const headers = {
+    "Content-Type": "application/json",
+  };
+
   const body = JSON.parse(event.body);
-  // Not too much data validation going on, if any further data validation errors it will just return an error
   if (body === null || body.id === null) {
     return {
       statusCode: 400,
-      headers: {
-        "Content-Type": "application/json",
-      },
+      headers: headers,
       body: JSON.stringify({
         data: "invalid request body",
       }),
@@ -21,32 +25,53 @@ exports.handler = async (event) => {
   }
 
   const { TABLE_NAME } = process.env;
-
   const client = new DynamoDBClient({ region: "us-east-1" });
-  const command = new DeleteItemCommand({
+
+  let command = new GetItemCommand({
     TableName: TABLE_NAME,
     Key: {
       Id: { S: body.id },
     },
-    ReturnValues: "ALL_OLD",
   });
   try {
-    const result = await client.send(command);
-    return {
-      statusCode: result.$metadata.httpStatusCode,
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        data: result.Attributes,
-      }),
-    };
+    let result = await client.send(command);
+    if (result.Item) {
+      console.log(result.Item);
+      if (result.Item.UserId.S == userId) {
+        const command = new DeleteItemCommand({
+          TableName: TABLE_NAME,
+          Key: {
+            Id: { S: body.id },
+          },
+          ReturnValues: "ALL_OLD",
+        });
+        const result = await client.send(command);
+        return {
+          statusCode: result.$metadata.httpStatusCode,
+          headers: headers,
+          body: JSON.stringify({
+            data: result.Attributes,
+          }),
+        };
+      } else {
+        return {
+          statusCode: 403,
+          headers: headers,
+        };
+      }
+    } else {
+      return {
+        statusCode: 404,
+        headers: headers,
+        body: JSON.stringify({
+          error: "invalid id",
+        }),
+      };
+    }
   } catch (err) {
     return {
       statusCode: 500,
-      headers: {
-        "Content-Type": "application/json",
-      },
+      headers: headers,
       body: JSON.stringify({
         error: err.message,
       }),
